# target/classes/META-INF/openapi.yml
openapi: 3.0.3
info:
  title: API de Produtos de Empréstimo - CAIXA (Desafio CAIXAVERSO)
  description: >
    API para CRUD de produtos de empréstimo e simulação de empréstimos usando
    taxa anual e sistema de amortização Price (parcelas fixas).
    Taxa efetiva mensal calculada por: i = (1 + i_anual)^(1/12) - 1.
  version: "1.0.0"
servers:
  - url: http://localhost:8080

tags:
  - name: Produtos
    description: Operações de CRUD sobre produtos de empréstimo
  - name: Simulação
    description: Endpoints de simulação de financiamento

paths:
  /produtos:
    get:
      tags:
        - Produtos
      summary: Listar produtos de empréstimo
      responses:
        "200":
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
    post:
      tags:
        - Produtos
      summary: Criar novo produto de empréstimo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarProdutoRequest'
      responses:
        "201":
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        "400":
          $ref: '#/components/responses/BadRequest'
  /produtos/{id}:
    get:
      tags:
        - Produtos
      summary: Obter produto por id
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        "200":
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Produtos
      summary: Atualizar produto
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarProdutoRequest'
      responses:
        "200":
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Produtos
      summary: Remover produto
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        "204":
          description: Produto removido
        "404":
          $ref: '#/components/responses/NotFound'

  /simulacoes/{idProduto}:
    post:
      tags:
        - Simulação
      summary: Simular financiamento para um produto
      description: |
        Recebe o produto (por id), valor solicitado e prazo em meses.
        Calcula taxa efetiva mensal: i = (1 + i_anual)^(1/12) - 1.
        Sistema de amortização: Price (parcelas fixas).
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulacaoRequest'
      responses:
        "200":
          description: Resultado da simulação (detalhamento mês a mês)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulacaoResponse'
        "422":
          description: Regras de negócio inválidas (ex. prazo > prazoMaximo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'


components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: ID do produto

  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Produto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: "Empréstimo Pessoal"
        taxaJurosAnual:
          type: number
          format: double
          description: Taxa anual em porcentagem (ex. 18.0)
          example: 18.0
        prazoMaximoMeses:
          type: integer
          example: 24
      required:
        - id
        - nome
        - taxaJurosAnual
        - prazoMaximoMeses

    CriarProdutoRequest:
      type: object
      properties:
        nome:
          type: string
        taxaJurosAnual:
          type: number
          format: double
          description: Taxa anual em porcentagem (ex. 18.0)
        prazoMaximoMeses:
          type: integer
      required:
        - nome
        - taxaJurosAnual
        - prazoMaximoMeses

    AtualizarProdutoRequest:
      type: object
      properties:
        nome:
          type: string
        taxaJurosAnual:
          type: number
          format: double
        prazoMaximoMeses:
          type: integer
      required:
        - nome
        - taxaJurosAnual
        - prazoMaximoMeses

    SimulacaoRequest:
      type: object
      properties:
        valorSolicitado:
          type: number
          format: double
          example: 10000.00
        prazoMeses:
          type: integer
          example: 24
      required:
        - valorSolicitado
        - prazoMeses

    MemoriaCalculo:
      type: object
      properties:
        mes:
          type: integer
          example: 1
        saldoDevedorInicial:
          type: number
          format: double
          example: 10000.00
        juros:
          type: number
          format: double
          example: 138.88
        amortizacao:
          type: number
          format: double
          example: 353.94
        parcela:
          type: number
          format: double
          example: 492.82
        saldoDevedorFinal:
          type: number
          format: double
          example: 9646.06
      required:
        - mes
        - saldoDevedorInicial
        - juros
        - amortizacao
        - parcela
        - saldoDevedorFinal

    SimulacaoResponse:
      type: object
      properties:
        produto:
          $ref: '#/components/schemas/Produto'
        valorSolicitado:
          type: number
          format: double
          example: 10000.00
        prazoMeses:
          type: integer
          example: 24
        taxaJurosAnual:
          type: number
          format: double
          example: 18.0
        taxaJurosEfetivaMensal:
          type: number
          format: double
          description: Taxa efetiva mensal (i = (1 + i_anual)^(1/12) - 1)
          example: 0.0138884303
        parcelaMensal:
          type: number
          format: double
          example: 492.82
        valorTotalComJuros:
          type: number
          format: double
          example: 11827.68
        memoriaCalculo:
          type: array
          items:
            $ref: '#/components/schemas/MemoriaCalculo'
      required:
        - produto
        - valorSolicitado
        - prazoMeses
        - taxaJurosAnual
        - taxaJurosEfetivaMensal
        - parcelaMensal
        - valorTotalComJuros
        - memoriaCalculo

    Error:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Prazo solicitado excede o prazo máximo do produto (24 meses)."

  securitySchemes:
    # placeholder (opcional) - se desejar JWT no futuro, podemos ativar facilmente
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT